Improve point calculation precision for user sessions
Update API to round points to one decimal place, adjust app to parse floats, and migrate database columns to support decimal values.

+++ artifacts/ryzer-app/app/session/result.tsx
@@ -41,7 +41,7 @@ export default function ResultScreen() {
   const [multiplierApplied, setMultiplierApplied] = useState(false);
   const [currentStreak, setCurrentStreak] = useState(0);
 
-  const ptNum = parseInt(points ?? "0");
+  const ptNum = parseFloat(points ?? "0");
   const durNum = parseInt(durationSeconds ?? "0");
 
   useEffect(() => {
